pyproject.toml
[project]
name = "dict-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aerich==0.9.1",
    "aiosqlite==0.21.0",
    "annotated-types==0.7.0",
    "anyio==4.10.0",
    "async-timeout==5.0.1",
    "asyncclick==8.2.2.2",
    "asyncmy==0.2.10",
    "bcrypt==4.3.0",
    "certifi==2025.8.3",
    "cffi==1.17.1",
    "charset-normalizer==3.4.3",
    "click==8.2.1",
    "colorama==0.4.6",
    "cryptography==45.0.6",
    "deprecated==1.2.18",
    "dictdiffer==0.9.0",
    "ecdsa==0.19.1",
    "et-xmlfile==2.0.0",
    "fastapi==0.116.1",
    "fugashi==1.5.1",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "idna==3.10",
    "iso8601==2.1.0",
    "jaconv==0.4.0",
    "numpy==2.3.2",
    "openpyxl==3.1.5",
    "pandas==2.3.1",
    "pandas-stubs==2.3.2.250827",
    "pyasn1==0.6.1",
    "pycparser==2.22",
    "pycryptodome==3.23.0",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pykakasi==2.3.0",
    "pypika-tortoise==0.6.1",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.1",
    "python-jose==3.5.0",
    "pytz==2025.2",
    "redis==6.4.0",
    "requests==2.32.5",
    "rsa==4.9.1",
    "six==1.17.0",
    "sniffio==1.3.1",
    "starlette==0.47.2",
    "tortoise-orm==0.25.1",
    "typing-extensions==4.14.1",
    "typing-inspection==0.4.1",
    "tzdata==2025.2",
    "unidic-lite==1.0.8",
    "urllib3==2.5.0",
    "uvicorn==0.35.0",
    "wrapt==1.17.3",
]

[tool.aerich]
tortoise_orm = "settings.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."

[tool.uv.sources]
[tool.uv.sources.default]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
